requirements.txt
gradio
python-dotenv
langchain
openai
elevenlabs
numpy
git+https://github.com/suno-ai/bark.git#egg=suno-bark
transformers
bitsandbytes
torch
accelerate
# document loaders
pypdf
playwright
unstructured
chromadb
tiktoken
sentence_transformers
xformers

# image generate
git+https://github.com/huggingface/diffusers#egg=diffusers
omegaconf

# sdxl refiner
#invisible_watermark
safetensors
